Migrate int use to float use re tank level

+++ com/components/DefaultLevelStrategy.py
@@ -6,5 +6,5 @@ class DefaultLevelStrategy(LevelStrategy):
     def __init__(self, name: str = "default level strategy"):
         super().__init__(name)
 
-    def get_level(self) -> int:
+    def get_level(self) -> float:
         pass
